+++ examples/szengine/a_header_szengine.py
@@ -6,18 +6,14 @@ Simply a header used in development.
 
 from typing import List, Tuple
 
+from senzing import SZ_WITHOUT_INFO, SzEngine
 from senzing_truthset import (
     TRUTHSET_CUSTOMER_RECORDS,
     TRUTHSET_REFERENCE_RECORDS,
     TRUTHSET_WATCHLIST_RECORDS,
 )
 
-from senzing_core import (
-    SZ_WITHOUT_INFO,
-    SzAbstractFactory,
-    SzAbstractFactoryParameters,
-    SzEngine,
-)
+from senzing_core import SzAbstractFactory, SzAbstractFactoryParameters
 
 DATA_SOURCES = {
     "CUSTOMERS": TRUTHSET_CUSTOMER_RECORDS,