+++ src/senzing_core/_helpers.py
@@ -31,7 +31,7 @@ from functools import wraps
 from types import TracebackType
 from typing import Any, Callable, Dict, List, Optional, Type, TypeVar, Union
 
-from senzing_abstract import ENGINE_EXCEPTION_MAP
+from senzing import ENGINE_EXCEPTION_MAP
 
 from senzing_core import SzError