+++ tests/szconfig_test.py
@@ -3,17 +3,16 @@ from typing import Any, Dict
 
 import pytest
 from pytest_schema import Optional, Or, schema
+from senzing import SzConfig as SzConfigAbstract
 
-from senzing_core import SzConfig
-from senzing_core import SzConfigCore as SzConfigTest
-from senzing_core import SzConfigurationError
+from senzing_core import SzConfig, SzConfigurationError
 
 # -----------------------------------------------------------------------------
 # Testcases
 # -----------------------------------------------------------------------------
 
 
-def test_add_data_source(sz_config: SzConfigTest) -> None:
+def test_add_data_source(sz_config: SzConfig) -> None:
     """Test SzConfig().add_data_source()."""
     data_source_code = "NAME_OF_DATASOURCE"
     config_handle = sz_config.create_config()
@@ -24,7 +23,7 @@ def test_add_data_source(sz_config: SzConfigTest) -> None:
     assert schema(add_data_source_schema) == actual_as_dict
 
 
-def test_add_data_source_bad_config_handle_type(sz_config: SzConfigTest) -> None:
+def test_add_data_source_bad_config_handle_type(sz_config: SzConfig) -> None:
     """Test SzConfig().add_data_source()."""
     bad_config_handle = "string"
     data_source_code = "NAME_OF_DATASOURCE"
@@ -32,7 +31,7 @@ def test_add_data_source_bad_config_handle_type(sz_config: SzConfigTest) -> None
         sz_config.add_data_source(bad_config_handle, data_source_code)  # type: ignore[arg-type]
 
 
-def test_add_data_source_bad_data_source_code_type(sz_config: SzConfigTest) -> None:
+def test_add_data_source_bad_data_source_code_type(sz_config: SzConfig) -> None:
     """Test SzConfig().add_data_source()."""
     config_handle = sz_config.create_config()
     bad_data_source_code = 0
@@ -43,7 +42,7 @@ def test_add_data_source_bad_data_source_code_type(sz_config: SzConfigTest) -> N
         sz_config.close_config(config_handle)
 
 
-def test_add_data_source_bad_data_source_code_value(sz_config: SzConfigTest) -> None:
+def test_add_data_source_bad_data_source_code_value(sz_config: SzConfig) -> None:
     """Test SzConfig().add_data_source()."""
     config_handle = sz_config.create_config()
     bad_data_source_code = {"XXXX": "YYYY"}
@@ -54,14 +53,14 @@ def test_add_data_source_bad_data_source_code_value(sz_config: SzConfigTest) -> 
         sz_config.close_config(config_handle)
 
 
-def test_close_config_bad_config_handle_type(sz_config: SzConfigTest) -> None:
+def test_close_config_bad_config_handle_type(sz_config: SzConfig) -> None:
     """Test SzConfig().create()."""
     bad_config_handle = "string"
     with pytest.raises(TypeError):
         sz_config.close_config(bad_config_handle)  # type: ignore[arg-type]
 
 
-def test_create_config(sz_config: SzConfigTest) -> None:
+def test_create_config(sz_config: SzConfig) -> None:
     """Test SzConfig().create()."""
     config_handle = sz_config.create_config()
     assert isinstance(config_handle, int)
@@ -71,7 +70,7 @@ def test_create_config(sz_config: SzConfigTest) -> None:
     assert config_handle > 0
 
 
-def test_delete_data_source(sz_config: SzConfigTest) -> None:
+def test_delete_data_source(sz_config: SzConfig) -> None:
     """Test SzConfig().delete_data_source()."""
     data_source_code = "TEST"
     config_handle = sz_config.create_config()
@@ -79,7 +78,7 @@ def test_delete_data_source(sz_config: SzConfigTest) -> None:
     sz_config.close_config(config_handle)
 
 
-def test_delete_data_source_bad_config_handle_type(sz_config: SzConfigTest) -> None:
+def test_delete_data_source_bad_config_handle_type(sz_config: SzConfig) -> None:
     """Test SzConfig().delete_data_source()."""
     data_source_code = "TEST"
     bad_config_handle = "string"
@@ -87,7 +86,7 @@ def test_delete_data_source_bad_config_handle_type(sz_config: SzConfigTest) -> N
         sz_config.delete_data_source(bad_config_handle, data_source_code)  # type: ignore[arg-type]
 
 
-def test_delete_data_source_bad_data_source_code_type(sz_config: SzConfigTest) -> None:
+def test_delete_data_source_bad_data_source_code_type(sz_config: SzConfig) -> None:
     """Test SzConfig().delete_data_source()."""
     bad_data_source_code = 0
     config_handle = sz_config.create_config()
@@ -96,7 +95,7 @@ def test_delete_data_source_bad_data_source_code_type(sz_config: SzConfigTest) -
     sz_config.close_config(config_handle)
 
 
-def test_delete_data_source_bad_data_source_code_value(sz_config: SzConfigTest) -> None:
+def test_delete_data_source_bad_data_source_code_value(sz_config: SzConfig) -> None:
     """Test SzConfig().delete_data_source()."""
     bad_data_source_code = {"XXXX": "YYYY"}
     config_handle = sz_config.create_config()
@@ -105,7 +104,7 @@ def test_delete_data_source_bad_data_source_code_value(sz_config: SzConfigTest) 
     sz_config.close_config(config_handle)
 
 
-def test_get_data_sources(sz_config: SzConfigTest) -> None:
+def test_get_data_sources(sz_config: SzConfig) -> None:
     """Test SzConfig().get_data_sources()."""
     config_handle = sz_config.create_config()
     actual = sz_config.get_data_sources(config_handle)
@@ -115,14 +114,14 @@ def test_get_data_sources(sz_config: SzConfigTest) -> None:
     assert schema(get_data_sources_schema) == actual_as_dict
 
 
-def test_get_data_sources_bad_config_handle_type(sz_config: SzConfigTest) -> None:
+def test_get_data_sources_bad_config_handle_type(sz_config: SzConfig) -> None:
     """Test SzConfig().list_data_sources()."""
     bad_config_handle = "string"
     with pytest.raises(TypeError):
         sz_config.get_data_sources(bad_config_handle)  # type: ignore[arg-type]
 
 
-def test_import_config(sz_config: SzConfigTest) -> None:
+def test_import_config(sz_config: SzConfig) -> None:
     """Test SzConfig().import_config()."""
     config_handle = sz_config.create_config()
     config_definition = sz_config.export_config(config_handle)
@@ -132,21 +131,21 @@ def test_import_config(sz_config: SzConfigTest) -> None:
     sz_config.close_config(config_handle)
 
 
-def test_import_config_bad_config_definition_type(sz_config: SzConfigTest) -> None:
+def test_import_config_bad_config_definition_type(sz_config: SzConfig) -> None:
     """Test SzConfig().import_config()."""
     bad_config_definition = 0
     with pytest.raises(TypeError):
         sz_config.import_config(bad_config_definition)  # type: ignore[arg-type]
 
 
-def test_import_config_bad_config_definition_value(sz_config: SzConfigTest) -> None:
+def test_import_config_bad_config_definition_value(sz_config: SzConfig) -> None:
     """Test SzConfig().import_config()."""
     bad_config_definition = '{"Just": "Junk"}'
     with pytest.raises(SzConfigurationError):
         sz_config.import_config(bad_config_definition)
 
 
-def test_export_config(sz_config: SzConfigTest) -> None:
+def test_export_config(sz_config: SzConfig) -> None:
     """Test SzConfig().export_config()."""
     config_handle = sz_config.create_config()
     actual = sz_config.export_config(config_handle)
@@ -156,7 +155,7 @@ def test_export_config(sz_config: SzConfigTest) -> None:
     assert schema(export_config_schema) == actual_as_dict
 
 
-def test_export_config_bad_config_handle_type(sz_config: SzConfigTest) -> None:
+def test_export_config_bad_config_handle_type(sz_config: SzConfig) -> None:
     """Test SzConfig().export_config()."""
     bad_config_handle = "string"
     with pytest.raises(TypeError):
@@ -170,12 +169,12 @@ def test_export_config_bad_config_handle_type(sz_config: SzConfigTest) -> None:
 
 def test_constructor(engine_vars: Dict[Any, Any]) -> None:
     """Test constructor."""
-    actual = SzConfigTest()
+    actual = SzConfig()
     actual._initialize(  # pylint: disable=W0212
         engine_vars["INSTANCE_NAME"],
         engine_vars["SETTINGS"],
     )
-    assert isinstance(actual, SzConfig)
+    assert isinstance(actual, SzConfigAbstract)
 
 
 # def test_constructor_bad_instance_name(engine_vars: Dict[Any, Any]) -> None:
@@ -204,17 +203,17 @@ def test_constructor(engine_vars: Dict[Any, Any]) -> None:
 
 def test_constructor_dict(engine_vars: Dict[Any, Any]) -> None:
     """Test constructor."""
-    actual = SzConfigTest()
+    actual = SzConfig()
     actual._initialize(  # pylint: disable=W0212
         engine_vars["INSTANCE_NAME"],
         engine_vars["SETTINGS_DICT"],
     )
-    assert isinstance(actual, SzConfig)
+    assert isinstance(actual, SzConfigAbstract)
 
 
 def test_destroy(engine_vars: Dict[Any, Any]) -> None:
     """Test constructor."""
-    actual = SzConfigTest()
+    actual = SzConfig()
     actual._initialize(  # pylint: disable=W0212
         engine_vars["INSTANCE_NAME"],
         engine_vars["SETTINGS"],
@@ -222,7 +221,7 @@ def test_destroy(engine_vars: Dict[Any, Any]) -> None:
     actual._destroy()  # pylint: disable=W0212
 
 
-def test_exception(sz_config: SzConfigTest) -> None:
+def test_exception(sz_config: SzConfig) -> None:
     """Test exceptions."""
     with pytest.raises(Exception):
         sz_config.check_result(-1)
@@ -234,13 +233,13 @@ def test_exception(sz_config: SzConfigTest) -> None:
 
 
 @pytest.fixture(name="sz_config", scope="function")
-def szconfig_fixture(engine_vars: Dict[Any, Any]) -> SzConfigTest:
+def szconfig_fixture(engine_vars: Dict[Any, Any]) -> SzConfig:
     """
     Single szconfig object to use for all tests.
     engine_vars is returned from conftest.py.
     """
 
-    result = SzConfigTest()
+    result = SzConfig()
     result._initialize(  # pylint: disable=W0212
         engine_vars["INSTANCE_NAME"],
         engine_vars["SETTINGS"],