+++ tests/szdiagnostic_test.py
@@ -3,19 +3,16 @@ from typing import Any, Dict
 
 import pytest
 from pytest_schema import schema
+from senzing import SzDiagnostic as SzDiagnosticAbstract
 
-from senzing_core import SzConfigManagerCore as SzConfigManagerTest
-from senzing_core import SzDiagnostic
-from senzing_core import SzDiagnosticCore as SzDiagnosticTest
-from senzing_core import SzEngineCore as SzEngineTest
-from senzing_core import SzError
+from senzing_core import SzConfigManager, SzDiagnostic, SzEngine, SzError
 
 # -----------------------------------------------------------------------------
 # Testcases
 # -----------------------------------------------------------------------------
 
 
-def test_check_datastore_performance(sz_diagnostic: SzDiagnosticTest) -> None:
+def test_check_datastore_performance(sz_diagnostic: SzDiagnostic) -> None:
     """Test SzDiagnostic().check_datastore_performance()."""
     seconds_to_run = 3
     actual = sz_diagnostic.check_datastore_performance(seconds_to_run)
@@ -24,7 +21,7 @@ def test_check_datastore_performance(sz_diagnostic: SzDiagnosticTest) -> None:
 
 
 def test_check_datastore_performance_bad_seconds_to_run_type(
-    sz_diagnostic: SzDiagnosticTest,
+    sz_diagnostic: SzDiagnostic,
 ) -> None:
     """Test SzDiagnostic().check_datastore_performance()."""
     bad_seconds_to_run = "string"
@@ -33,7 +30,7 @@ def test_check_datastore_performance_bad_seconds_to_run_type(
 
 
 def test_check_datastore_performance_bad_seconds_to_run_value(
-    sz_diagnostic: SzDiagnosticTest,
+    sz_diagnostic: SzDiagnostic,
 ) -> None:
     """Test SzDiagnostic().check_datastore_performance()."""
     bad_seconds_to_run = -1
@@ -44,14 +41,14 @@ def test_check_datastore_performance_bad_seconds_to_run_value(
     assert schema(check_datastore_performance_schema) == actual_as_dict
 
 
-def test_get_datastore_info(sz_diagnostic: SzDiagnosticTest) -> None:
+def test_get_datastore_info(sz_diagnostic: SzDiagnostic) -> None:
     """Test SzDiagnostic().get_datastore_info()."""
     actual = sz_diagnostic.get_datastore_info()
     actual_as_dict = json.loads(actual)
     assert schema(get_datastore_info_schema) == actual_as_dict
 
 
-def test_get_feature(sz_diagnostic: SzDiagnosticTest, sz_engine: SzEngineTest) -> None:
+def test_get_feature(sz_diagnostic: SzDiagnostic, sz_engine: SzEngine) -> None:
     """# TODO"""
     data_source_code = "TEST"
     record_id = "1"
@@ -62,7 +59,7 @@ def test_get_feature(sz_diagnostic: SzDiagnosticTest, sz_engine: SzEngineTest) -
     assert schema(get_feature_schema) == actual_as_dict
 
 
-def test_get_feature_unknown_id(sz_diagnostic: SzDiagnosticTest) -> None:
+def test_get_feature_unknown_id(sz_diagnostic: SzDiagnostic) -> None:
     """# TODO"""
     with pytest.raises(SzError):
         _ = sz_diagnostic.get_feature(111111111111111111)
@@ -75,12 +72,12 @@ def test_get_feature_unknown_id(sz_diagnostic: SzDiagnosticTest) -> None:
 
 def test_constructor(engine_vars: Dict[Any, Any]) -> None:
     """Test constructor."""
-    actual = SzDiagnosticTest()
+    actual = SzDiagnostic()
     actual._initialize(  # pylint: disable=W0212
         engine_vars["INSTANCE_NAME"],
         engine_vars["SETTINGS"],
     )
-    assert isinstance(actual, SzDiagnostic)
+    assert isinstance(actual, SzDiagnosticAbstract)
 
 
 # def test_constructor_bad_instance_name(engine_vars: Dict[Any, Any]) -> None:
@@ -109,17 +106,17 @@ def test_constructor(engine_vars: Dict[Any, Any]) -> None:
 
 def test_constructor_dict(engine_vars: Dict[Any, Any]) -> None:
     """Test constructor."""
-    actual = SzDiagnosticTest()
+    actual = SzDiagnostic()
     actual._initialize(  # pylint: disable=W0212
         engine_vars["INSTANCE_NAME"],
         engine_vars["SETTINGS_DICT"],
     )
-    assert isinstance(actual, SzDiagnostic)
+    assert isinstance(actual, SzDiagnosticAbstract)
 
 
 def test_destroy(engine_vars: Dict[Any, Any]) -> None:
     """Test constructor."""
-    actual = SzDiagnosticTest()
+    actual = SzDiagnostic()
     actual._initialize(  # pylint: disable=W0212
         engine_vars["INSTANCE_NAME"],
         engine_vars["SETTINGS"],
@@ -127,13 +124,13 @@ def test_destroy(engine_vars: Dict[Any, Any]) -> None:
     actual._destroy()
 
 
-def test_exception(sz_configmanager: SzConfigManagerTest) -> None:
+def test_exception(sz_configmanager: SzConfigManager) -> None:
     """Test exceptions."""
     with pytest.raises(Exception):
         sz_configmanager.check_result(-1)
 
 
-def test_reinitialize(sz_diagnostic: SzDiagnosticTest, sz_configmanager: SzConfigManagerTest) -> None:
+def test_reinitialize(sz_diagnostic: SzDiagnostic, sz_configmanager: SzConfigManager) -> None:
     """Test SzDiagnostic().reinit() with current config ID."""
     default_config_id = sz_configmanager.get_default_config_id()
     try:
@@ -142,14 +139,14 @@ def test_reinitialize(sz_diagnostic: SzDiagnosticTest, sz_configmanager: SzConfi
         assert False
 
 
-def test_reinitialize_bad_config_id(sz_diagnostic: SzDiagnosticTest) -> None:
+def test_reinitialize_bad_config_id(sz_diagnostic: SzDiagnostic) -> None:
     """Test SzDiagnostic().reinit() with current config ID."""
     bad_default_config_id = "string"
     with pytest.raises(TypeError):
         sz_diagnostic._reinitialize(bad_default_config_id)  # type: ignore[arg-type]
 
 
-def test_reinitialize_missing_config_id(sz_diagnostic: SzDiagnosticTest) -> None:
+def test_reinitialize_missing_config_id(sz_diagnostic: SzDiagnostic) -> None:
     """Test SzDiagnostic().reinit() raising error."""
     with pytest.raises(SzError):
         sz_diagnostic._reinitialize(999)
@@ -161,10 +158,10 @@ def test_reinitialize_missing_config_id(sz_diagnostic: SzDiagnosticTest) -> None
 
 
 @pytest.fixture(name="sz_configmanager", scope="function")
-def szconfigmanager_fixture(engine_vars: Dict[Any, Any]) -> SzConfigManagerTest:
+def szconfigmanager_fixture(engine_vars: Dict[Any, Any]) -> SzConfigManager:
     """Single szconfigmanager object to use for all tests.
     engine_vars is returned from conftest.pys"""
-    result = SzConfigManagerTest()
+    result = SzConfigManager()
     result._initialize(
         instance_name=engine_vars["INSTANCE_NAME"],
         settings=engine_vars["SETTINGS"],
@@ -174,10 +171,10 @@ def szconfigmanager_fixture(engine_vars: Dict[Any, Any]) -> SzConfigManagerTest:
 
 
 @pytest.fixture(name="sz_diagnostic", scope="function")
-def szdiagnostic_fixture(engine_vars: Dict[Any, Any]) -> SzDiagnosticTest:
+def szdiagnostic_fixture(engine_vars: Dict[Any, Any]) -> SzDiagnostic:
     """Single szdiagnostic object to use for all tests.
     engine_vars is returned from conftest.pys"""
-    result = SzDiagnosticTest()
+    result = SzDiagnostic()
     result._initialize(
         instance_name=engine_vars["INSTANCE_NAME"],
         settings=engine_vars["SETTINGS"],
@@ -188,10 +185,10 @@ def szdiagnostic_fixture(engine_vars: Dict[Any, Any]) -> SzDiagnosticTest:
 
 
 @pytest.fixture(name="sz_engine", scope="function")
-def szengine_fixture(engine_vars: Dict[Any, Any]) -> SzEngineTest:
+def szengine_fixture(engine_vars: Dict[Any, Any]) -> SzEngine:
     """Single szengine object to use for all tests.
     engine_vars is returned from conftest.pys"""
-    result = SzEngineTest()
+    result = SzEngine()
     result._initialize(
         instance_name=engine_vars["INSTANCE_NAME"],
         settings=engine_vars["SETTINGS"],